checkfac(made4)	R Documentation

Check factor

Description

Convert input vector into a factor suitable for analysis in ADE4. This function is called by link[made4:bga]{bga} and other made4 function, it is not normally called by the user

Usage

checkfac(classvec)

Arguments

classvec 	 A vector or factor

Details

Value

factor 	 factor suitable for input into bga

Note

Author(s)

Aedin Culhane

See Also

Examples

[Package Contents]